[swan] Move bell code to bell-swan.c, simplify Makefile/config changes
