+++ elks/arch/i86/drivers/char/config.in
@@ -15,10 +15,8 @@ mainmenu_option next_comment
 			choice 'Console font'			\
 				"8x8   CONFIG_CONSOLE_FONT_8X8  \
 				4x8    CONFIG_CONSOLE_FONT_4X8" 4x8
-			bool '  Console bell'    CONFIG_CONSOLE_BELL y
-		else
-			bool '  Scancode keyboard driver'	CONFIG_KEYBOARD_SCANCODE y
 		fi
+		bool '  Scancode keyboard driver'	CONFIG_KEYBOARD_SCANCODE y
 		bool '  Dual-screen console support'    CONFIG_CONSOLE_DUAL n
 	fi
 	bool 'Serial Console'	CONFIG_CONSOLE_SERIAL		n
@@ -26,9 +24,7 @@ mainmenu_option next_comment
 	        bool 'Console ANSI terminal emulation'  CONFIG_EMUL_ANSI  y
 	fi
 	if [ "$CONFIG_CONSOLE_DIRECT" = "y" ]; then
-		if [ "$CONFIG_ARCH_SWAN" = "n" ]; then
-			source elks/arch/i86/drivers/char/KeyMaps/config.in
-		fi
+		source elks/arch/i86/drivers/char/KeyMaps/config.in
 	fi
 
 	comment 'Other character devices'
@@ -37,9 +33,7 @@ mainmenu_option next_comment
 	else
 		define_bool CONFIG_CHAR_DEV_RS y
 	fi
-	if [ "$CONFIG_ARCH_SWAN" = "n" ]; then
-		bool 'Parallel device driver'	CONFIG_CHAR_DEV_LP	  y
-	fi
+	bool 'Parallel device driver'	CONFIG_CHAR_DEV_LP	  y
 	#bool 'CGA text mode videobuffer'	CONFIG_CHAR_DEV_CGATEXT	  n
 	bool 'Memory character devices'		CONFIG_CHAR_DEV_MEM	  y
 	#if [ "$CONFIG_CHAR_DEV_MEM" = "y" ]; then
