+++ swan.config
@@ -38,7 +38,7 @@ CONFIG_ROMCODE=y
 CONFIG_ROM_SETUP_CODE=f000
 CONFIG_ROM_KERNEL_CODE=f060
 CONFIG_ROM_SETUP_DATA=0040
-CONFIG_ROM_KERNEL_DATA=0060
+CONFIG_ROM_KERNEL_DATA=0061
 CONFIG_ROM_BASE=f000
 CONFIG_ROM_CHECKSUM_SIZE=64
 CONFIG_ROM_BOOTABLE_BY_RESET=y
@@ -103,7 +103,6 @@ CONFIG_CONSOLE_DIRECT=y
 # CONFIG_CONSOLE_HEADLESS is not set
 # CONFIG_CONSOLE_FONT_8X8 is not set
 CONFIG_CONSOLE_FONT_4X8=y
-# CONFIG_CONSOLE_BELL is not set
 # CONFIG_CONSOLE_DUAL is not set
 # CONFIG_CONSOLE_SERIAL is not set
 CONFIG_EMUL_ANSI=y
